reorganize gustav forms and user scripts
