+++ gustav/forms/html/__init__.py
@@ -20,3 +20,4 @@
 # Comments and/or additions are welcome. Send e-mail to: cbrown1@pitt.edu.
 #
 
+from . import nafc
